+++ CHANGELOG.md
@@ -10,6 +10,7 @@ title: Changelog
 * `<cc-html-frame>`:
   * call `revokeObjectURL` when the component is disconnected
   * add `loading` feature
+* `<cc-invoice>`: use `loading` feature from `<cc-html-frame>`
 
 ## 6.3.0 (2021-04-13)